backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
firebase-admin==6.6.0
python-dotenv==1.0.1
python-multipart==0.0.20
websockets==14.1
pydantic==2.10.4
